Removed bad casts.

+++ src/main/java/com/orsoncharts/graphics3d/Point3D.java
@@ -61,9 +61,9 @@ public final class Point3D implements Serializable {
      * @param z  the z-coordinate.
      */
     public Point3D(double x, double y, double z) {
-        this.x = (float) x;
-        this.y = (float) y;
-        this.z = (float) z;
+        this.x = x;
+        this.y = y;
+        this.z = z;
     }
     
     /**
